Extract url querying function into a helper class
